+++ changelog.md
@@ -13,5 +13,5 @@
    - Remove fullname-derived-namespace = namespace-in-type-attributes assertition (#60)
    - Add avro_ocf:make_ocf/2 to allow building ocf content in memory
 * 2.6.0
-   - Support Parsing Canonical Form for Schemas (contributer @congini @reachfh)
+   - Support Parsing Canonical Form for Schemas (contributer congini/@reachfh)